feat: avoid hard coded sdk package path in protocol package
